+++ src/main/java/at/hannibal2/skyhanni/features/garden/fortuneguide/FFStats.kt
@@ -75,7 +75,6 @@ object FFStats {
             CropType.SUNFLOWER, CropType.MOONFLOWER, CropType.WILD_ROSE,
             -> {
                 FortuneStats.BASE_TOOL.set(FarmingFortuneDisplay.getToolFortune(tool), 50.0)
-                FortuneStats.COUNTER.set(FarmingFortuneDisplay.getCounterFortune(tool), 96.0)
                 FortuneStats.HARVESTING.set(FarmingFortuneDisplay.getHarvestingFortune(tool), 75.0)
                 FortuneStats.COLLECTION.set(FarmingFortuneDisplay.getCollectionFortune(tool), 48.0)
                 FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 20.0)