+++ src/main/java/at/hannibal2/skyhanni/features/rift/area/wyldwoods/ShyCruxWarnings.kt
@@ -7,14 +7,14 @@ import at.hannibal2.skyhanni.data.title.TitleManager
 import at.hannibal2.skyhanni.features.rift.RiftApi
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.EntityUtils
-import at.hannibal2.skyhanni.utils.LocationUtils.distanceToPlayer
+import net.minecraft.entity.Entity
 import kotlin.time.Duration.Companion.milliseconds
 
 @SkyHanniModule
 object ShyCruxWarnings {
 
     private val config get() = RiftApi.config.area.wyldWoods
-    private val shyNames = arrayOf("I'm ugly! :(", "Eek!", "Don't look at me!", "Look away!")
+    private val shyNames = setOf("I'm ugly! :(", "Eek!", "Don't look at me!", "Look away!")
 
     @HandleEvent(onlyOnIsland = IslandType.THE_RIFT)
     fun onTick() {
@@ -23,7 +23,7 @@ object ShyCruxWarnings {
     }
 
     private fun checkForShy() {
-        if (EntityUtils.getAllEntities().any { it.name in shyNames && it.distanceToPlayer() < 8 }) {
+        if (EntityUtils.getEntitiesNextToPlayer<Entity>(8.0).any { it.name in shyNames }) {
             TitleManager.sendTitle("§eLook away!", duration = 150.milliseconds)
         }
     }