+++ docs/CHANGELOG.md
@@ -57,6 +57,7 @@
 + Hide Mythological Creatures Since text for creatures you haven't gotten. - Empa (https://github.com/hannibal002/SkyHanni/pull/4861)
 + Improved Diana burrow guessing. - SidOfThe7Cs (https://github.com/hannibal002/SkyHanni/pull/4916)
   + Added an option to instantly detect the next burrow from the arrow.
++ Only Show Coa counter during diana. - nopo (https://github.com/hannibal002/SkyHanni/pull/4942)
 
 #### Hunting
 
@@ -71,6 +72,8 @@
   + Made the solution also allow swapping end columns as apparently you can do that.
   + Made it work if you abiphone Jake.
   + Blocked accidentally clicking the close puzzle button.
++ Added support for Moonflower, Sunflower, and Wild Rose in Basket of Seeds for Not Clickable Items. - Luna (https://github.com/hannibal002/SkyHanni/pull/4954)
++ Added chips to visitor drop rewards display. - nopo (https://github.com/hannibal002/SkyHanni/pull/4960)
 
 #### Mining
 
@@ -105,6 +108,7 @@
   + Only enabled on sessions by default; uptime was not tracked before this and will be wrong on total mode.
 + Made Block Break Prticle Hider searchable by typing breaking. - Luna (https://github.com/hannibal002/SkyHanni/pull/4931)
 + Made warnings about missing tab list info less frequent and aggressive. - Luna (https://github.com/hannibal002/SkyHanni/pull/4933)
++ Improved performance of entity features. - Empa (https://github.com/hannibal002/SkyHanni/pull/4900)
 
 ### Fixes
 
@@ -150,11 +154,22 @@
 + Fixed Slayer spawn costs not being tracked during aura. - nopo (https://github.com/hannibal002/SkyHanni/pull/4905)
 + Fixed Broodmother chat messages sending twice or incorrectly. - MTOnline (https://github.com/hannibal002/SkyHanni/pull/4918)
 + Fixed Slayer Spawn Cost not Counting during Aura. - fazfoxy (https://github.com/hannibal002/SkyHanni/pull/4919)
++ Fixed Fire Pillar Display disappearing sometimes wrongly. - Empa (https://github.com/hannibal002/SkyHanni/pull/4900)
 
 #### Farming
 
 + Fixed Burrowing Spores notification not working. - Luna (https://github.com/hannibal002/SkyHanni/pull/4883)
 + Fixed Fungi Cutter Warning not working on 1.21. - Luna (https://github.com/hannibal002/SkyHanni/pull/4896)
++ Fixed composter overlay not working sometimes (hopefully). - nopo (https://github.com/hannibal002/SkyHanni/pull/4944)
++ Fixed error when getting a tool capsule. - nopo (https://github.com/hannibal002/SkyHanni/pull/4943)
++ Fixed Hoe Level Display not saving its location. - nopo (https://github.com/hannibal002/SkyHanni/pull/4958)
++ Fixed Jacob's Contest spam around SB new year. - nopo (https://github.com/hannibal002/SkyHanni/pull/4955)
++ Fixed Nether Wart Fortune stat not being properly detected on items. - Luna (https://github.com/hannibal002/SkyHanni/pull/4947)
++ Fixed Organic matter display going negative. - nopo (https://github.com/hannibal002/SkyHanni/pull/4957)
++ Fixed Skill error during crop fever. - nopo (https://github.com/hannibal002/SkyHanni/pull/4950)
++ Fixed vinyl error in the vacuum. - nopo (https://github.com/hannibal002/SkyHanni/pull/4949)
++ Made armor drop tracker work with the greenhouse update. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4953)
++ Fixed sowdust displaying incorrectly in chunked stats on custom scoreboard. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4963)
 
 #### Chocolate Factory
 
@@ -187,6 +202,11 @@
 + Fixed glowing invisible mobs on 1.21.10. - nopo (https://github.com/hannibal002/SkyHanni/pull/4913)
 + Fixed error when opening Your Skills menu with Roman numerals disabled in Hypixel settings. - Luna (https://github.com/hannibal002/SkyHanni/pull/4932)
 + Fixed overflow features not having support for hannibal numerals. - nopo (https://github.com/hannibal002/SkyHanni/pull/4935)
++ Fixed error when getting a trophy fish. - nopo (https://github.com/hannibal002/SkyHanni/pull/4943)
++ Fixed Hideonleaf finder navigating when typing. - nopo (https://github.com/hannibal002/SkyHanni/pull/4945)
++ Fixed some box rendering on modern. - nopo (https://github.com/hannibal002/SkyHanni/pull/4940)
++ Fixed Tablist not rendering over other elements when using toggle tab. - nopo (https://github.com/hannibal002/SkyHanni/pull/4941)
++ Fixed potential config wipe. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4961)
 
 ### Technical Details
 
@@ -209,12 +229,18 @@
   + They can still be edited via `/shconfig`.
 + Added the ability to change current mayor perks via repo. - nopo (https://github.com/hannibal002/SkyHanni/pull/4936)
 + Using `@OnlyLegacy` and `@OnlyModern` together on the same config option now throws a compile-time (and runtime) error. - Luna (https://github.com/hannibal002/SkyHanni/pull/4908)
++ Added the ability to remove something from the config. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4961)
++ Made config moves actually remove the old data. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4961)
 
 ### Removed Features
 
 + Removed Fix NEU Heavy Pearls. - nopo (https://github.com/hannibal002/SkyHanni/pull/4881)
   + The bug in NEU was fixed long time ago.
 + Removed the Tool Tooltip Tweaks feature. - nopo (https://github.com/hannibal002/SkyHanni/pull/4926)
++ Removed dicer rng tracker. - CalMWolfs (https://github.com/hannibal002/SkyHanni/pull/4953)
+  + Dicers no longer drop crops like that.
++ Removed Fungi Cutter Warning. - Luna (https://github.com/hannibal002/SkyHanni/pull/4951)
+  + Fungi Cutters no longer have a mode.
 
 ## Version 5.0.0